+++ saleor/graphql/checkout/mutations.py
@@ -41,6 +41,7 @@ from ...product.models import ProductChannelListing
 from ...shipping import models as shipping_models
 from ...warehouse import models as warehouse_models
 from ...warehouse.availability import check_stock_and_preorder_quantity_bulk
+from ...warehouse.reservations import get_reservation_length, is_reservation_enabled
 from ..account.i18n import I18nMixin
 from ..account.types import AddressInput
 from ..channel.utils import clean_channel
@@ -128,6 +129,7 @@ def check_lines_quantity(
     allow_zero_quantity=False,
     existing_lines=None,
     replace=False,
+    check_reservations=False,
 ):
     """Clean quantities and check if stock is sufficient for each checkout line.
 
@@ -176,6 +178,7 @@ def check_lines_quantity(
             channel_slug,
             existing_lines=existing_lines,
             replace=replace,
+            check_reservations=check_reservations,
         )
     except InsufficientStock as e:
         errors = [
@@ -290,7 +293,7 @@ class CheckoutCreate(ModelMutation, I18nMixin):
 
     @classmethod
     def clean_checkout_lines(
-        cls, lines, country, channel
+        cls, info, lines, country, channel
     ) -> Tuple[List[product_models.ProductVariant], List[int]]:
         variant_ids = [line["variant_id"] for line in lines]
         variants = cls.get_nodes_or_error(
@@ -308,7 +311,13 @@ class CheckoutCreate(ModelMutation, I18nMixin):
         validate_variants_available_in_channel(
             variant_db_ids, channel.id, CheckoutErrorCode.UNAVAILABLE_VARIANT_IN_CHANNEL
         )
-        check_lines_quantity(variants, quantities, country, channel.slug)
+        check_lines_quantity(
+            variants,
+            quantities,
+            country,
+            channel.slug,
+            check_reservations=is_reservation_enabled(info.context.site.settings),
+        )
         return variants, quantities
 
     @classmethod
@@ -354,7 +363,12 @@ class CheckoutCreate(ModelMutation, I18nMixin):
             (
                 cleaned_input["variants"],
                 cleaned_input["quantities"],
-            ) = cls.clean_checkout_lines(lines, country, cleaned_input["channel"])
+            ) = cls.clean_checkout_lines(
+                info,
+                lines,
+                country,
+                cleaned_input["channel"],
+            )
 
         # Use authenticated user's email as default email
         if user.is_authenticated:
@@ -385,7 +399,13 @@ class CheckoutCreate(ModelMutation, I18nMixin):
         quantities = cleaned_input.get("quantities")
         if variants and quantities:
             try:
-                add_variants_to_checkout(instance, variants, quantities, channel.slug)
+                add_variants_to_checkout(
+                    instance,
+                    variants,
+                    quantities,
+                    channel.slug,
+                    reservation_length=get_reservation_length(info.context),
+                )
             except InsufficientStock as exc:
                 error = prepare_insufficient_stock_checkout_validation_error(exc)
                 raise ValidationError({"lines": error})
@@ -458,15 +478,21 @@ class CheckoutLinesAdd(BaseMutation):
 
     @classmethod
     def validate_checkout_lines(
-        cls, variants, quantities, country, channel_slug, lines=None
+        cls, info, variants, quantities, country, channel_slug, lines=None
     ):
         check_lines_quantity(
-            variants, quantities, country, channel_slug, existing_lines=lines
+            variants,
+            quantities,
+            country,
+            channel_slug,
+            existing_lines=lines,
+            check_reservations=is_reservation_enabled(info.context.site.settings),
         )
 
     @classmethod
     def clean_input(
         cls,
+        info,
         checkout,
         variants,
         quantities,
@@ -477,8 +503,14 @@ class CheckoutLinesAdd(BaseMutation):
         replace,
     ):
         channel_slug = checkout_info.channel.slug
+
         cls.validate_checkout_lines(
-            variants, quantities, checkout.get_country(), channel_slug, lines=lines
+            info,
+            variants,
+            quantities,
+            checkout.get_country(),
+            channel_slug,
+            lines=lines,
         )
         variants_db_ids = {variant.id for variant in variants}
         validate_variants_available_for_purchase(variants_db_ids, checkout.channel_id)
@@ -497,6 +529,8 @@ class CheckoutLinesAdd(BaseMutation):
                     channel_slug,
                     skip_stock_check=True,  # already checked by validate_checkout_lines
                     replace=replace,
+                    replace_reservations=True,
+                    reservation_length=get_reservation_length(info.context),
                 )
             except ProductNotPublished as exc:
                 raise ValidationError(
@@ -545,6 +579,7 @@ class CheckoutLinesAdd(BaseMutation):
 
         lines = fetch_checkout_lines(checkout)
         lines = cls.clean_input(
+            info,
             checkout,
             variants,
             quantities,
@@ -584,7 +619,13 @@ class CheckoutLinesUpdate(CheckoutLinesAdd):
 
     @classmethod
     def validate_checkout_lines(
-        cls, variants, quantities, country, channel_slug, lines=None
+        cls,
+        info,
+        variants,
+        quantities,
+        country,
+        channel_slug,
+        lines=None,
     ):
         check_lines_quantity(
             variants,
@@ -594,6 +635,7 @@ class CheckoutLinesUpdate(CheckoutLinesAdd):
             allow_zero_quantity=True,
             existing_lines=lines,
             replace=True,
+            check_reservations=is_reservation_enabled(info.context.site.settings),
         )
 
     @classmethod
@@ -794,7 +836,11 @@ class CheckoutShippingAddressUpdate(BaseMutation, I18nMixin):
 
     @classmethod
     def process_checkout_lines(
-        cls, lines: Iterable["CheckoutLineInfo"], country: str, channel_slug: str
+        cls,
+        info,
+        lines: Iterable["CheckoutLineInfo"],
+        country: str,
+        channel_slug: str,
     ) -> None:
         variant_ids = [line_info.variant.id for line_info in lines]
         variants = list(
@@ -803,7 +849,17 @@ class CheckoutShippingAddressUpdate(BaseMutation, I18nMixin):
             ).prefetch_related("product__product_type")
         )  # FIXME: is this prefetch needed?
         quantities = [line_info.line.quantity for line_info in lines]
-        check_lines_quantity(variants, quantities, country, channel_slug)
+        check_lines_quantity(
+            variants,
+            quantities,
+            country,
+            channel_slug,
+            # Set replace=True to avoid existing_lines and quantities from
+            # being counted twice by the check_stock_quantity_bulk
+            replace=True,
+            existing_lines=lines,
+            check_reservations=is_reservation_enabled(info.context.site.settings),
+        )
 
     @classmethod
     def perform_mutation(
@@ -864,7 +920,7 @@ class CheckoutShippingAddressUpdate(BaseMutation, I18nMixin):
 
         # Resolve and process the lines, validating variants quantities
         if lines:
-            cls.process_checkout_lines(lines, country, checkout_info.channel.slug)
+            cls.process_checkout_lines(info, lines, country, checkout_info.channel.slug)
 
         update_checkout_shipping_method_if_invalid(checkout_info, lines)