+++ saleor/graphql/checkout/tests/test_checkout_lines.py
@@ -1,16 +1,21 @@
 import datetime
 import uuid
+from decimal import Decimal
 from unittest import mock
 
 import graphene
 import pytest
+from django.utils import timezone
 
 from ....checkout.error_codes import CheckoutErrorCode
 from ....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from ....checkout.utils import calculate_checkout_quantity
+from ....checkout.models import Checkout
+from ....checkout.utils import add_variant_to_checkout, calculate_checkout_quantity
 from ....plugins.manager import get_plugins_manager
 from ....product.models import ProductChannelListing
 from ....warehouse import WarehouseClickAndCollectOption
+from ....warehouse.models import Reservation, Stock
+from ....warehouse.tests.utils import get_available_quantity_for_stock
 from ...tests.utils import get_graphql_content
 from ..mutations import update_checkout_shipping_method_if_invalid
 
@@ -67,11 +72,140 @@ def test_checkout_lines_add(
     assert line.variant == variant
     assert line.quantity == 1
     assert calculate_checkout_quantity(lines) == 4
+    assert not Reservation.objects.exists()
 
     manager = get_plugins_manager()
     lines = fetch_checkout_lines(checkout)
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
     mocked_update_shipping_method.assert_called_once_with(checkout_info, lines)
+
+
+def test_checkout_lines_add_with_reservations(
+    site_settings_with_reservations, user_api_client, checkout_with_item, stock
+):
+    variant = stock.product_variant
+    checkout = checkout_with_item
+    line = checkout.lines.first()
+    lines = fetch_checkout_lines(checkout)
+    assert calculate_checkout_quantity(lines) == 3
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_id, "quantity": 1}],
+        "channelSlug": checkout.channel.slug,
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["checkoutLinesAdd"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    line = checkout.lines.latest("pk")
+    assert line.variant == variant
+    assert line.quantity == 1
+    assert calculate_checkout_quantity(lines) == 4
+
+    reservation = line.reservations.get()
+    assert reservation
+    assert reservation.checkout_line == line
+    assert reservation.quantity_reserved == line.quantity
+
+
+def test_checkout_lines_add_updates_reservation(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_line_with_one_reservation,
+    stock,
+):
+    variant = checkout_line_with_one_reservation.variant
+    checkout = checkout_line_with_one_reservation.checkout
+    line = checkout_line_with_one_reservation
+    lines = fetch_checkout_lines(checkout)
+    assert calculate_checkout_quantity(lines) == 2
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_id, "quantity": 1}],
+        "channelSlug": checkout.channel.slug,
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["checkoutLinesAdd"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    line = checkout.lines.latest("pk")
+    assert line.variant == variant
+    assert line.quantity == 3
+    assert calculate_checkout_quantity(lines) == 3
+
+    reservation = line.reservations.get()
+    assert reservation
+    assert reservation.checkout_line == line
+    assert reservation.quantity_reserved == line.quantity
+
+
+def test_checkout_lines_add_new_variant_updates_other_lines_reservations_expirations(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_line_with_one_reservation,
+    product,
+    warehouse,
+):
+    variant = checkout_line_with_one_reservation.variant
+    checkout = checkout_line_with_one_reservation.checkout
+    line = checkout_line_with_one_reservation
+    reservation = line.reservations.get()
+    lines = fetch_checkout_lines(checkout)
+    assert calculate_checkout_quantity(lines) == 2
+
+    variant_other = product.variants.create(sku="SKU_B")
+    variant_other.channel_listings.create(
+        channel=checkout.channel,
+        price_amount=Decimal(10),
+        cost_price_amount=Decimal(1),
+        currency=checkout.channel.currency_code,
+    )
+    Stock.objects.create(
+        product_variant=variant_other, warehouse=warehouse, quantity=15
+    )
+    variant_other_id = graphene.Node.to_global_id("ProductVariant", variant_other.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_other_id, "quantity": 3}],
+        "channelSlug": checkout.channel.slug,
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["checkoutLinesAdd"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    line.refresh_from_db()
+    assert line.variant == variant
+    assert line.quantity == 2
+    new_line = checkout.lines.get(variant=variant_other)
+    assert new_line.variant == variant_other
+    assert new_line.quantity == 3
+    assert calculate_checkout_quantity(lines) == 5
+
+    updated_reservation = Reservation.objects.get(checkout_line__variant=variant)
+    assert updated_reservation.checkout_line == line
+    assert updated_reservation.quantity_reserved == line.quantity
+    assert updated_reservation.reserved_until > reservation.reserved_until
+
+    other_reservation = Reservation.objects.get(checkout_line__variant=variant_other)
+    assert other_reservation.checkout_line == new_line
+    assert other_reservation.quantity_reserved == new_line.quantity
+    assert other_reservation.reserved_until > reservation.reserved_until
+
+    assert updated_reservation.reserved_until == other_reservation.reserved_until
+
+    with pytest.raises(Reservation.DoesNotExist):
+        reservation.refresh_from_db()
 
 
 def test_checkout_lines_add_existing_variant(user_api_client, checkout_with_item):
@@ -147,6 +281,42 @@ def test_checkout_lines_add_with_insufficient_stock(
     variables = {
         "token": checkout.token,
         "lines": [{"variantId": variant_id, "quantity": 49}],
+        "channelSlug": checkout.channel.slug,
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
+    content = get_graphql_content(response)
+    errors = content["data"]["checkoutLinesAdd"]["errors"]
+    assert errors[0]["code"] == CheckoutErrorCode.INSUFFICIENT_STOCK.name
+    assert errors[0]["field"] == "quantity"
+
+
+def test_checkout_lines_add_with_reserved_insufficient_stock(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_with_item,
+    stock,
+    channel_USD,
+):
+    variant = stock.product_variant
+    checkout = checkout_with_item
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
+    quantity_available = get_available_quantity_for_stock(stock)
+
+    other_checkout = Checkout.objects.create(channel=channel_USD, currency="USD")
+    other_checkout_line = other_checkout.lines.create(
+        variant=variant,
+        quantity=quantity_available - 1,
+    )
+    Reservation.objects.create(
+        checkout_line=other_checkout_line,
+        stock=stock,
+        quantity_reserved=quantity_available - 1,
+        reserved_until=timezone.now() + datetime.timedelta(minutes=5),
+    )
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_id, "quantity": 2}],
         "channelSlug": checkout.channel.slug,
     }
     response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
@@ -475,6 +645,164 @@ def test_checkout_lines_update(
     mocked_update_shipping_method.assert_called_once_with(checkout_info, lines)
 
 
+def test_checkout_lines_update_with_new_reservations(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_line_with_reservation_in_many_stocks,
+):
+    assert Reservation.objects.count() == 2
+    checkout = checkout_line_with_reservation_in_many_stocks.checkout
+    lines = fetch_checkout_lines(checkout)
+    assert checkout.lines.count() == 1
+    assert calculate_checkout_quantity(lines) == 3
+    line = checkout.lines.first()
+    variant = line.variant
+    assert line.quantity == 3
+
+    Stock.objects.filter(
+        warehouse__shipping_zones__countries__contains="US", product_variant=variant
+    ).update(quantity=3)
+
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_id, "quantity": 1}],
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_UPDATE, variables)
+    content = get_graphql_content(response)
+
+    data = content["data"]["checkoutLinesUpdate"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    assert checkout.lines.count() == 1
+    line = checkout.lines.first()
+    assert line.variant == variant
+    assert line.quantity == 1
+    assert calculate_checkout_quantity(lines) == 1
+
+    reservation = checkout_line_with_reservation_in_many_stocks.reservations.first()
+    assert reservation.quantity_reserved == line.quantity
+
+    assert Reservation.objects.count() == 1
+
+
+def test_checkout_lines_update_against_reserved_stock(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_line,
+    stock,
+    channel_USD,
+):
+    assert Reservation.objects.count() == 0
+    checkout = checkout_line.checkout
+    lines = fetch_checkout_lines(checkout)
+    assert checkout.lines.count() == 1
+    assert calculate_checkout_quantity(lines) == 3
+    variant = checkout_line.variant
+    assert checkout_line.quantity == 3
+
+    other_checkout = Checkout.objects.create(channel=channel_USD, currency="USD")
+    other_checkout_line = other_checkout.lines.create(
+        variant=variant,
+        quantity=7,
+    )
+    reservation = Reservation.objects.create(
+        checkout_line=other_checkout_line,
+        stock=variant.stocks.first(),
+        quantity_reserved=7,
+        reserved_until=timezone.now() + datetime.timedelta(minutes=5),
+    )
+
+    variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_id, "quantity": 5}],
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_UPDATE, variables)
+    content = get_graphql_content(response)
+
+    data = content["data"]["checkoutLinesUpdate"]
+    assert data["errors"]
+    assert data["errors"][0]["message"] == (
+        "Could not add items 123. Only 3 remaining in stock."
+    )
+    assert data["errors"][0]["field"] == "quantity"
+
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    assert checkout.lines.count() == 1
+    line = checkout.lines.first()
+    assert line.variant == variant
+    assert line.quantity == 3
+    assert calculate_checkout_quantity(lines) == 3
+    reservation.refresh_from_db()
+    assert Reservation.objects.count() == 1
+
+
+def test_checkout_lines_update_other_lines_reservations_expirations(
+    site_settings_with_reservations,
+    user_api_client,
+    checkout_line_with_one_reservation,
+    product,
+    warehouse,
+):
+    variant = checkout_line_with_one_reservation.variant
+    checkout = checkout_line_with_one_reservation.checkout
+    line = checkout_line_with_one_reservation
+    reservation = line.reservations.get()
+    checkout_info = fetch_checkout_info(checkout, [], [], get_plugins_manager())
+    lines = fetch_checkout_lines(checkout)
+    assert calculate_checkout_quantity(lines) == 2
+
+    variant_other = product.variants.create(sku="SKU_B")
+    variant_other.channel_listings.create(
+        channel=checkout.channel,
+        price_amount=Decimal(10),
+        cost_price_amount=Decimal(1),
+        currency=checkout.channel.currency_code,
+    )
+    Stock.objects.create(
+        product_variant=variant_other, warehouse=warehouse, quantity=15
+    )
+    add_variant_to_checkout(checkout_info, variant_other, 2)
+    variant_other_id = graphene.Node.to_global_id("ProductVariant", variant_other.pk)
+
+    variables = {
+        "token": checkout.token,
+        "lines": [{"variantId": variant_other_id, "quantity": 3}],
+    }
+    response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["checkoutLinesUpdate"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    lines = fetch_checkout_lines(checkout)
+    line.refresh_from_db()
+    assert line.variant == variant
+    assert line.quantity == 2
+    new_line = checkout.lines.get(variant=variant_other)
+    assert new_line.variant == variant_other
+    assert new_line.quantity == 3
+    assert calculate_checkout_quantity(lines) == 5
+
+    updated_reservation = Reservation.objects.get(checkout_line__variant=variant)
+    assert updated_reservation.checkout_line == line
+    assert updated_reservation.quantity_reserved == line.quantity
+    assert updated_reservation.reserved_until > reservation.reserved_until
+
+    other_reservation = Reservation.objects.get(checkout_line__variant=variant_other)
+    assert other_reservation.checkout_line == new_line
+    assert other_reservation.quantity_reserved == new_line.quantity
+
+    assert updated_reservation.reserved_until == other_reservation.reserved_until
+
+    with pytest.raises(Reservation.DoesNotExist):
+        reservation.refresh_from_db()
+
+
 def test_checkout_lines_update_with_unavailable_variant(
     user_api_client, checkout_with_item
 ):
@@ -739,9 +1067,12 @@ MUTATION_CHECKOUT_LINES_DELETE = """
     wraps=update_checkout_shipping_method_if_invalid,
 )
 def test_checkout_line_delete(
-    mocked_update_shipping_method, user_api_client, checkout_with_item
+    mocked_update_shipping_method,
+    user_api_client,
+    checkout_line_with_reservation_in_many_stocks,
 ):
-    checkout = checkout_with_item
+    assert Reservation.objects.count() == 2
+    checkout = checkout_line_with_reservation_in_many_stocks.checkout
     lines = fetch_checkout_lines(checkout)
     assert calculate_checkout_quantity(lines) == 3
     assert checkout.lines.count() == 1
@@ -760,6 +1091,7 @@ def test_checkout_line_delete(
     lines = fetch_checkout_lines(checkout)
     assert checkout.lines.count() == 0
     assert calculate_checkout_quantity(lines) == 0
+    assert Reservation.objects.count() == 0
     manager = get_plugins_manager()
     checkout_info = fetch_checkout_info(checkout, lines, [], manager)
     mocked_update_shipping_method.assert_called_once_with(checkout_info, lines)