+++ saleor/graphql/shop/mutations.py
@@ -60,6 +60,18 @@ class ShopSettingsInput(graphene.InputObjectType):
     customer_set_password_url = graphene.String(
         description="URL of a view where customers can set their password."
     )
+    reserve_stock_duration_anonymous_user = graphene.Int(
+        description=(
+            f"{ADDED_IN_31} Default number of minutes stock will be reserved for "
+            "anonymous checkout. Enter 0 or null to disable."
+        )
+    )
+    reserve_stock_duration_authenticated_user = graphene.Int(
+        description=(
+            f"{ADDED_IN_31} Default number of minutes stock will be reserved for "
+            "authenticated checkout. Enter 0 or null to disable."
+        )
+    )
 
 
 class SiteDomainInput(graphene.InputObjectType):
@@ -90,6 +102,16 @@ class ShopSettingsUpdate(BaseMutation):
                 raise ValidationError(
                     {"customer_set_password_url": error}, code=ShopErrorCode.INVALID
                 )
+
+        if "reserve_stock_duration_anonymous_user" in data:
+            new_value = data["reserve_stock_duration_anonymous_user"]
+            if not new_value or new_value < 1:
+                data["reserve_stock_duration_anonymous_user"] = None
+        if "reserve_stock_duration_authenticated_user" in data:
+            new_value = data["reserve_stock_duration_authenticated_user"]
+            if not new_value or new_value < 1:
+                data["reserve_stock_duration_authenticated_user"] = None
+
         return data
 
     @classmethod