+++ saleor/site/models.py
@@ -70,6 +70,12 @@ class SiteSettings(models.Model):
     fulfillment_auto_approve = models.BooleanField(default=True)
     fulfillment_allow_unpaid = models.BooleanField(default=True)
 
+    # Duration in minutes
+    reserve_stock_duration_anonymous_user = models.IntegerField(blank=True, null=True)
+    reserve_stock_duration_authenticated_user = models.IntegerField(
+        blank=True, null=True
+    )
+
     # gift card settings
     gift_card_expiry_type = models.CharField(
         max_length=32,