+++ saleor/warehouse/availability.py
@@ -6,13 +6,19 @@ from django.db.models.functions import Coalesce
 
 from ..core.exceptions import InsufficientStock, InsufficientStockData
 from ..product.models import ProductVariantChannelListing
-from .models import Stock, StockQuerySet
+from .models import Reservation, Stock, StockQuerySet
 
 if TYPE_CHECKING:
-    from ..product.models import ProductVariant
+    from ..checkout.fetch import CheckoutLineInfo
+    from ..checkout.models import CheckoutLine
+    from ..product.models import Product, ProductVariant
 
 
-def _get_available_quantity(stocks: StockQuerySet) -> int:
+def _get_available_quantity(
+    stocks: StockQuerySet,
+    checkout_lines: Optional[List["CheckoutLine"]] = None,
+    check_reservations: bool = False,
+) -> int:
     results = stocks.aggregate(
         total_quantity=Coalesce(Sum("quantity", distinct=True), 0),
         quantity_allocated=Coalesce(Sum("allocations__quantity_allocated"), 0),
@@ -20,7 +26,12 @@ def _get_available_quantity(stocks: StockQuerySet) -> int:
     total_quantity = results["total_quantity"]
     quantity_allocated = results["quantity_allocated"]
 
-    return max(total_quantity - quantity_allocated, 0)
+    if check_reservations:
+        quantity_reserved = get_reserved_quantity(stocks, checkout_lines)
+    else:
+        quantity_reserved = 0
+
+    return max(total_quantity - quantity_allocated - quantity_reserved, 0)
 
 
 def check_stock_and_preorder_quantity(
@@ -38,7 +49,12 @@ def check_stock_and_preorder_quantity(
 
 
 def check_stock_quantity(
-    variant: "ProductVariant", country_code: str, channel_slug: str, quantity: int
+    variant: "ProductVariant",
+    country_code: str,
+    channel_slug: str,
+    quantity: int,
+    checkout_lines: Optional[List["CheckoutLine"]] = None,
+    check_reservations: bool = False,
 ):
     """Validate if there is stock available for given variant in given country.
 
@@ -52,7 +68,10 @@ def check_stock_quantity(
         if not stocks:
             raise InsufficientStock([InsufficientStockData(variant=variant)])
 
-        if quantity > _get_available_quantity(stocks):
+        available_quantity = _get_available_quantity(
+            stocks, checkout_lines, check_reservations
+        )
+        if quantity > available_quantity:
             raise InsufficientStock([InsufficientStockData(variant=variant)])
 
 
@@ -64,6 +83,7 @@ def check_stock_and_preorder_quantity_bulk(
     additional_filter_lookup: Optional[Dict[str, Any]] = None,
     existing_lines: Iterable = None,
     replace: bool = False,
+    check_reservations: bool = False,
 ):
     """Validate if products are available for stocks/preorder.
 
@@ -85,6 +105,7 @@ def check_stock_and_preorder_quantity_bulk(
             additional_filter_lookup,
             existing_lines,
             replace,
+            check_reservations,
         )
     if preorder_variants:
         check_preorder_threshold_bulk(
@@ -122,8 +143,9 @@ def check_stock_quantity_bulk(
     quantities: Iterable[int],
     channel_slug: str,
     additional_filter_lookup: Optional[Dict[str, Any]] = None,
-    existing_lines: Iterable = None,
+    existing_lines: Iterable["CheckoutLineInfo"] = None,
     replace=False,
+    check_reservations: bool = False,
 ):
     """Validate if there is stock available for given variants in given country.
 
@@ -143,6 +165,14 @@ def check_stock_quantity_bulk(
     for stock in all_variants_stocks:
         variant_stocks[stock.product_variant_id].append(stock)
 
+    if check_reservations:
+        variant_reservations = get_reserved_quantity_bulk(
+            all_variants_stocks,
+            [line.line for line in existing_lines] if existing_lines else [],
+        )
+    else:
+        variant_reservations = defaultdict(int)
+
     insufficient_stocks: List[InsufficientStockData] = []
     variants_quantities = {
         line.variant.pk: line.line.quantity for line in existing_lines or []
@@ -154,6 +184,9 @@ def check_stock_quantity_bulk(
         stocks = variant_stocks.get(variant.pk, [])
         available_quantity = sum(
             [stock.available_quantity for stock in stocks]  # type: ignore
+        )
+        available_quantity = max(
+            available_quantity - variant_reservations[variant.pk], 0
         )
 
         if quantity > 0:
@@ -240,3 +273,75 @@ def check_preorder_threshold_bulk(
 
     if insufficient_stocks:
         raise InsufficientStock(insufficient_stocks)
+
+
+def get_available_quantity(
+    variant: "ProductVariant",
+    country_code: str,
+    channel_slug: str,
+    checkout_lines: Optional[List["CheckoutLine"]] = None,
+    check_reservations: bool = False,
+) -> int:
+    """Return available quantity for given product in given country."""
+    stocks = Stock.objects.get_variant_stocks_for_country(
+        country_code, channel_slug, variant
+    )
+    if not stocks:
+        return 0
+    return _get_available_quantity(stocks, checkout_lines, check_reservations)
+
+
+def is_product_in_stock(
+    product: "Product", country_code: str, channel_slug: str
+) -> bool:
+    """Check if there is any variant of given product available in given country."""
+    stocks = Stock.objects.get_product_stocks_for_country_and_channel(
+        country_code, channel_slug, product
+    ).annotate_available_quantity()
+    return any(stocks.values_list("available_quantity", flat=True))
+
+
+def get_reserved_quantity(
+    stocks: StockQuerySet, lines: Optional[List["CheckoutLine"]] = None
+) -> int:
+    result = (
+        Reservation.objects.filter(
+            stock__in=stocks,
+        )
+        .not_expired()
+        .exclude_checkout_lines(lines)
+        .aggregate(
+            quantity_reserved=Coalesce(Sum("quantity_reserved"), 0),
+        )
+    )  # type: ignore
+
+    return result["quantity_reserved"]
+
+
+def get_reserved_quantity_bulk(
+    stocks: Iterable[Stock],
+    checkout_lines: Iterable["CheckoutLine"],
+) -> Dict[int, int]:
+    reservations: Dict[int, int] = defaultdict(int)
+    if not stocks:
+        return reservations
+
+    result = (
+        Reservation.objects.filter(
+            stock__in=stocks,
+        )
+        .not_expired()
+        .exclude_checkout_lines(checkout_lines)
+        .values("stock_id")
+        .annotate(
+            quantity_reserved=Coalesce(Sum("quantity_reserved"), 0),
+        )
+    )  # type: ignore
+
+    stocks_variants = {stock.id: stock.product_variant_id for stock in stocks}
+    for stock_reservations in result:
+        variant_id = stocks_variants.get(stock_reservations["stock_id"])
+        if variant_id:
+            reservations[variant_id] += stock_reservations["quantity_reserved"]
+
+    return reservations